Fix build errors due to the -Wc++11-narrowing compilation flag
Since now this flag is being passed to the compiler, we can't use
the list-initializing syntax to build types when any of the parameters
specified for constructing the new object would need to be implicitly
converted into a different type of smaller size than the original one.

For instance, we can't now list initialize kPageTransitionQualifierMask
in the |PageTransitionType| enumeration to 0xFFFFFF00:

enum PageTransitionType : int32_t {
  [...]
  // General mask defining the bits used for the qualifiers.
  kPageTransitionQualifierMask = 0xFFFFFF00,  // THIS WILL FAIL NOW
};

...which would give us the following error at build time:

  enumerator value evaluates to 4294967040, which cannot be narrowed
      to type 'int32_t' (aka 'int') [-Wc++11-narrowing]

This is because 0xFFFFFF00 can't be fitted inside a 32-bit unsigned int
as the highest bit is reserved for the sign. Instead we could either
define the enumeration as of type uint32_t or, even better and probably
good enough in this case, to use 0x7FFFFF00 for the mask instead:

enum PageTransitionType : int32_t {
  [...]
  // General mask defining the bits used for the qualifiers.
  kPageTransitionQualifierMask = 0x7FFFFF00,  // THIS WORKS
};

The right solution depends on each case, which is why this patch fixes
the few problems found differently, in some cases by simply silencing
the compilers using a static_cast<T>(value), in others by avoiding
to use the list-initializing syntax, calling the constructor instead,
and in some others adjusting the type used to a more appropriate one.